requirements-test.txt
app_version>=0.1.1
django-appconf
pillow
django-override-settings
PyYAML
